Graphviz image file path support.
Provides environment variables for each Graphviz rendering process to look up image file names in a list of known locations:

1. User overrides: the working directory +/graphviz/images, +/graphviz, or +/.
2. Built-in stylesheet images: sketch path + /styles/images, +/styles, or +/.

An image filename referenced in a styles configuration file with no path specified will be found by the png image renderer if it is in any of these locations, or in a subdirectory specified relative to one of these locations.

Tested on Mac, not yet tested on Windows.