Depth ruler prototype
Rendered using GraphStream diameter, which does not correspond to the graphviz ranksep columns width -- it may be shorter or longer. This may mean that issue #17 (Depth meter for graphviz output) is not achievable with this kind of layout -- nodes will often appear in ranks that do not correspond to their shallowest depth, and due to loopbacks or merging the longest path may be longer than the deepest node column.